text, images, cure model

+++ src/Texture2d.cpp
@@ -1,11 +1,11 @@
-#include "SharedHeaders.h"
+#include "utils/SharedHeaders.h"
 #include "Texture2d.h"
 
 using std::string;
 using std::vector;
 
 namespace br {
-	Texture2d::Texture2d(string path, vector<uint8_t>& data, int16_t width, int16_t height)
+	Texture2d::Texture2d(string path, vector<uint8_t>& data, uint32_t width, uint32_t height)
 		: path(path), data(data), width(width), height(height) {
 	}
 }
